requirements.txt
langchain
langchain-community
openai
chromadb
sentence-transformers
pdfplumber
python-dotenv
requests
fastapi
uvicorn
tqdm